refactor(user): 타입 분리
타입 분리
