+++ types/user.ts
@@ -1,0 +1,8 @@
+export type UserPublicProfile = {
+	id: number;
+	teamId: string;
+	email: string;
+	name: string;
+	companyName?: string;
+	image: string | null;
+};
